requirements.txt
requests~=2.27.1
Pillow~=9.0.1
python-dotenv~=0.21.0